requirements-dev.txt
# Packages dependencies
celery>=5.0
python-dateutil
redis>=3.2
tenacity
backports.zoneinfo>=0.2.1; python_version < "3.9.0"
pytz

# Linting
black
flake8
flake8-black
flake8-isort

# Testing
fakeredis>=1.0.3
